disco-backend-scrape/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
lxml==5.3.0
cachetools==5.3.3
rapidfuzz==3.9.6
